flash/nor: add Nordic nRF53 support
Add flash support for the Nordic nRF53. The nRF53 bears some resemblance
to the nRF52, but the addition of a second core as well as the extra
security features make it difficult to share code between the two.
Therefore, nRF53 support is added as a separate driver.

Basic flashing operations are supported and tested. The code is designed
to handle various possible security configurations, but this has not
been thouroughly tested.

Flash protection support is not included in this patch. The nRF53
supports protecting the flash through the SPU and ACL peripherals, but
these are both volatile mechanisms, so OpenOCD would probably only
support reading the current protection state, and this did not seem like
a high priority.